Auto-adjust ifaddr= if set to the network address
It's very easy to accidentally put something like 172.16.0.0/16 of
fd42::/64 into the ifaddr setting, which isn't technically valid, so if
we are given the subnet 0 address, increment it by one and warn.

Also adds a check for IPv4 that we are not specifying the broadcast
address.  (IPv6 does not have a broadcast address).