Add rc fetch hashes to nodedb
Ideally we don't want clients fetching every RC frequently, even though
it is currently only a couple hundred kilobytes.  With enough clients
and enough frequency, that's still an annoying amount of data transfer,
and wasted processing.  To that end, this commit adds a bucketed hashing
system for RCs:

- Relays are divided into 128 buckets based on the 17th (arbitrary) byte
  of their RouterID.  A later byte is chosen to avoid clustering for
  nodes with "vanity" public keys, though perhaps clustering those would
  be a good thing for this?
- Serialized RCs are hashed (*except* for the timestamp), and these
  hashes are cached.
- Each bucket has a "hash" which is simply the XOR of all the RC hashes
  which belong to that bucket.  XOR is fine here because there is no
  cryptographic concern and it makes insertion/removal trivial.

This is all so that a client can request all RCs and provide *its*
current hash for each bucket, and the relay can respond with all the RCs
for *that* bucket, without the client needing to refresh the full RC
set.

The timestamp is ignored so that RC updates which are basically saying
"yes I'm still here, but nothing has changed" don't need to be pushed to
clients.  If an RC expires, the client will purge it as expired, then
its hash for that bucket will mismatch and it will get the newest RC for
that node.

This update mechanism is as-yet unimplemented; this commit simply adds
the hash buckets.