Only IPv6 by default; pubkey-based IPv6 addrs; fix A/AAAA DNS responses
This removes the default mapping of IPv4 addresses so that most internal
connections now speak only IPv6.  The only real use case we have for
IPv4 internally is for carrying IPv4 exit traffic, which we will revisit
and enable only when enabling exit mode.

This changes the IPv6 allocation to try assigning an IP based on the
pubkey of the remote first, so that, for instance, testnet decaf20
resolves to `fd2e:6c6f:6b69:0:deca:f200:25ca:6389` rather than
`fd2e:6c6f:6b69::2` (or whatever sequential value).  Sequential values
are still used as a fallback in case of pubkey prefix collision.

This also uncovered some bugs in how we respond to DNS queries:
- if requesting an A or AAAA record, we should only respond with that
  record type and not both the A/AAAA
- if requesting an A or AAAA and only an AAAA or A exists, we are
  supposed to return a NODATA reponse to indicate that the domain
  exists, but doesn't have the requested record type.  It doesn't seem
  as though we have ever done that properly, but now we do.
- The SERVFAIL response was commented out so if we didn't handle a
  request, DNS requests were just timing out.  This restores it.