Disconnect from deregged SNs after a delay
When a node leaves the network, currently we just keep an established
connection to that node alive as long as it stays up.

This fixes that by adding a check for deregged nodes and, if found,
disconnects them after half an hour.

The delay here is to attempt to not kill existing client paths that may
still be using a path through us to the deregged node, allowing some
time for the client to replace the path with new paths that don't use
it.