Properly close connections, and stop using close_quietly
Actually close the connections.

Quiet closing doesn't seem to have a purpose, and moreover seems like a
hack to avoid having to deal with proper callbacks, so this changes all
the "close quietly" to just "close".

This ended up triggering warnings in the close callback about closing an
untracked connection, so this also addresses those by using a specific
close error code for those connections so that we can detect them and
debug instead of warn for them.