Unbreak SNS; Add SNS response confirmations
SNS was broken but also not well designed: the SNS lookup path response
got double-encoded for no reason, but also didn't get properly
double-encoded, and it went through unnecessary supplemental objects
(within which the broken double-encoding occured).

This overhauls it to make it work properly: SNS path responses are now
single-encoded dicts with just nonce/ciphertext keys, and the extraneous
stuff is dropped.

This also adds a confirmation mechanism to SNS: instead of returning
when we get the first response, we now return once we get at least half
of all requests that we made that agree on the result.  (And if we don't
get that, we fall back to the most popular response of whatever we did
get).

Also removes a call to j.dump(); I hope this doesn't break anything!