Allow IPv4 for snodes, and on outbound paths
IPv4 will work fine for snodes, and might have a use case if you have a
client in some weird environment where IPv6 is disabled on the system.

For outbound client sessions, we need to be able to allow IPv4 because
an attempt to connect to a remote that already has an inbound session
with us will use the inbound session, even though that is an outbound
attempt.  What I mean is:

    - client A looks up ipv4.b.sesh, establishing a session with IPv4
      connectivity.
    - client B looks up ipv4.a.sesh.

The latter lookup *might* use an outbound or inbound path, depending on
the timing of A's session establishing to B, and so we need the other
end to support IPv4 allocation regardless of whether it was an inbound
or outbound.