Initiate CC refresh after remote close & re-add stale CC detection
This was removed in e311b2eddbba50d5a22c32b457260fa985fa69d2 but
something like it is needed after all: if we get a session close, we
need to start attempts to refetch the session so that we can reestablish
quickly (e.g. for the case that the remote restarted).  Without this
change, a remote that restarts resulted in 30s of known-dead Session
producing warnings before the nothing-for-30s timer kicks in and forces
a refresh.