libsession SNS resolution; add different "connected" states
This adds support for SNS resolution.

In doing so, I ran into a problem where the `on_connected` callback
would fire before we can actually resolve anything through SNS, and so
this also splits the "connected" status into two different concepts:
"edge connected" (which is what was there already), and a new concept of
"path connected", which means "edge connected" *and* an established
inbound path.  This adds the needed pieces to implement that concept
into Router.

Updates the libsession-router jank test to support ONS lookups (which
also needs the above for the required proper connectivity detection).