CI: install updates before oxen packages
We're going to install updates anyway; this just moves when they get
installed.

This fixes an issue on sid with the clang/llvm builds where a libc++-api
sub-package gets installed that has a conflict with a different
subpackage from an earlier version of the package that hadn't yet been
updated.