Add config options for common relay settings
Relay config generation commonly needs to set public ip, port, oxend rpc
address, and listen address.  Currently the multi-sn scripts used sed to
edit these into the generated file, but this update allows them to be
specified as part of the config generation so that the sed editing can
be dropped entirely.

This also puts them into a relay-specific options group (and adds a
client-specific options group) to make the --help output a little easier
to parse.