Redo tun auto-selection to retry up to 50 times
Currently we find a free IP address for the tun during configuration,
and then actually try to assign that much later in Router startup.

This can easily race with other session-router processes, causing them
to fail startup, and then have to get restarted (e.g. by systemd).

This is annoying, and this commit fixes it by moving the auto-selection
into the interface creation code itself, along with a retry loop in case
of conflicts to largely avoid such failed startups.  Essentially, now,
if the interface creation gets a 0 address, that means to auto-select a
range rather than us always passing in a specific range to try.

To deal with an issue of mapaddr config values getting processed during
config handling (where previously they were depending on the
auto-selected range), this now imposes a new config rule: any specified
mapaddrs require an explicit tun range and fail a new config check if
there is no explicit tun range, or if the mapped addresses are not
inside that tun range.